fix: move fixtureCleanup to testing helpers
