+++ packages/testing-helpers/fixture.d.ts
@@ -2,3 +2,4 @@ import { TemplateResult } from 'lit-html';
 
 export function fixtureSync(template: string | TemplateResult): Element;
 export function fixture(template: string | TemplateResult): Promise<Element>;
+export function fixtureCleanup(): void;
