+++ packages/testing-helpers/fixture.js
@@ -1,34 +1,16 @@
-import { TemplateResult } from 'lit-html';
-import { nextFrame } from './helpers.js';
-import { stringFixtureSync } from './stringFixture.js';
-import { litFixtureSync } from './litFixture.js';
+import { fixtureCleanup } from './fixtureWrapper.js';
+
+export { fixture, fixtureSync } from './fixture-no-side-effect.js';
 
 /**
- * Setups an element synchronously from the provided string template and puts it in the DOM.
- * Allows to specify properties via an object or a function taking the element as an argument.
- *
- * @param {string | TemplateResult} template
- * @returns {Element}
+ * This registers the fixture cleanup as a side effect
  */
-export function fixtureSync(template) {
-  if (typeof template === 'string') {
-    return stringFixtureSync(template);
+try {
+  if (afterEach) {
+    afterEach(() => {
+      fixtureCleanup();
+    });
   }
-  if (template instanceof TemplateResult) {
-    return litFixtureSync(template);
-  }
-  throw new Error('Invalid template provided - string or lit-html TemplateResult is supported');
-}
-
-/**
- * Setups an element asynchronously from the provided string template and puts it in the DOM.
- * Allows to specify properties via an object or a function taking the element as an argument.
- *
- * @param {string | TemplateResult} template
- * @returns {Promise<Element>}
- */
-export async function fixture(template) {
-  const result = fixtureSync(template);
-  await nextFrame();
-  return result;
+} catch (error) {
+  /* do nothing */
 }
