+++ packages/testing/index-no-side-effects.js
@@ -11,6 +11,7 @@ export {
   litFixtureSync,
   fixture,
   fixtureSync,
-} from '@open-wc/testing-helpers';
+  fixtureCleanup,
+} from '@open-wc/testing-helpers/index-no-side-effects.js';
 
 export { chai, expect, should, assert } from '@bundled-es-modules/chai';
