--- packages/testing/register-fixture-cleanup.js
@@ -1,12 +1,0 @@
-import { cachedWrappers } from '@open-wc/testing-helpers/fixtureWrapper.js';
-
-if (afterEach) {
-  afterEach(() => {
-    if (cachedWrappers) {
-      cachedWrappers.forEach(wrapper => {
-        document.body.removeChild(wrapper);
-      });
-    }
-    cachedWrappers.length = 0; // reset it like this as we can't reassign it
-  });
-}
