+++ packages/testing/test/side-effects.test.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-unused-expressions */
 
-import { cachedWrappers } from '@open-wc/testing-helpers/fixtureWrapper.js';
 import { fixture, expect } from '../index.js';
 
 describe('BDD', () => {
@@ -9,10 +8,6 @@ describe('BDD', () => {
     const el2 = await fixture('<div class="el2"></div>');
     expect(el).to.not.be.undefined;
     expect(el2).to.not.be.undefined;
-  });
-
-  it('will cleanup after each test', async () => {
-    expect(cachedWrappers.length).to.equal(0);
   });
 
   it('uses chai dom equals plugin', async () => {
